+++ golang/lab/lab7/fruit.go
@@ -5,22 +5,22 @@ import "fmt"
 type Fruit struct {
 	Name      string
 	Freshness string
-	Price     float32
+	Price     float64
 }
 
 func (f Fruit) GetInfo() {
 	fmt.Println("У нас есть в наличии", f.Name, "оно", f.Freshness, "и стоит", f.Price)
 }
 
-func (f Fruit) GetPrice() float32 {
+func (f Fruit) GetPrice() float64 {
 	return f.Price
 }
 
-func (f *Fruit) MakeDiscount(x float32) {
+func (f *Fruit) MakeDiscount(x float64) {
 	(*f).Price = (f.Price / 100) * (100 - x)
 }
 
-func (f *Fruit) ChangePrice(x float32) {
+func (f *Fruit) ChangePrice(x float64) {
 	(*f).Price = x
 }
